Revert "Feature/faster ad tls v3"
