--- stan/math/rev/core/init_chainablestack.hpp
@@ -1,14 +1,0 @@
-#ifndef STAN_MATH_REV_CORE_INIT_CHAINABLESTACK_HPP
-#define STAN_MATH_REV_CORE_INIT_CHAINABLESTACK_HPP
-
-#include <stan/math/rev/core/chainablestack.hpp>
-
-namespace stan {
-namespace math {
-namespace {
-
-ChainableStack global_stack_instance_init;
-}
-}  // namespace math
-}  // namespace stan
-#endif
